adding balance screen with mocks and little refactor on files
