+++ lib/shared/components/transaction_type_dropdown/transaction_type_dropdown_cubit.dart
@@ -1,6 +1,6 @@
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:hermes_app/shared/components/transaction_type_dropdown/transaction_type_dropdown_state.dart';
-import 'package:hermes_app/shared/models/unknown_error.dart';
+import 'package:hermes_app/shared/entities/unknown_error.dart';
 import 'package:hermes_app/shared/usecases/get_transaction_types_to_dropdown_use_case.dart';
 import 'package:hermes_app/shared/widgets/dropdown/dropdown_item.dart';
