+++ lib/shared/entities/category_model.dart
@@ -1,13 +1,13 @@
 import 'package:flutter/material.dart';
-import 'package:hermes_app/shared/entities/transaction_type.dart';
+import 'package:hermes_app/shared/entities/transaction_type_model.dart';
 import 'package:hermes_app/shared/entities/unmapped_entities/base_model.dart';
 
-class Category extends BaseModel<Category> {
+class CategoryModel extends BaseModel<CategoryModel> {
   final String name;
-  final TransactionType transactionType;
+  final TransactionTypeModel transactionType;
   final String icon;
   final Color color;
-  const Category({
+  const CategoryModel({
     required int? id,
     required this.name,
     required this.transactionType,
