+++ lib/shared/entities/failure_model.dart
