+++ lib/shared/entities/nullable_model.dart
