+++ lib/shared/entities/transaction_type_model.dart
@@ -1,8 +1,9 @@
 import 'package:hermes_app/shared/entities/unmapped_entities/base_model.dart';
 
-class TransactionType extends BaseModel<TransactionType> {
+class TransactionTypeModel extends BaseModel<TransactionTypeModel> {
   final String name;
-  const TransactionType({
+
+  const TransactionTypeModel({
     required int? id,
     required this.name,
   }) : super(id: id);
@@ -15,8 +16,8 @@ class TransactionType extends BaseModel<TransactionType> {
     };
   }
 
-  factory TransactionType.fromMap(Map<String, dynamic> map) {
-    return TransactionType(
+  factory TransactionTypeModel.fromMap(Map<String, dynamic> map) {
+    return TransactionTypeModel(
       id: map['id'] as int?,
       name: map['name'] as String,
     );
