+++ lib/shared/repositories/transaction_type_repository.dart
@@ -1,19 +1,19 @@
-import 'package:hermes_app/shared/entities/transaction_type.dart';
+import 'package:hermes_app/shared/entities/transaction_type_model.dart';
 
 import '../constants/tables.dart';
 import '../database/idatabase.dart';
 import 'base_repository.dart';
 
-class TransactionTypeRepository extends BaseRepository<TransactionType> {
+class TransactionTypeRepository extends BaseRepository<TransactionTypeModel> {
   final IDatabase _db;
 
   TransactionTypeRepository(this._db) : super(_db, Tables.transactionType);
 
-  Future<List<TransactionType>> getAll() async {
+  Future<List<TransactionTypeModel>> getAll() async {
     final result = await _db.getAll(Tables.transactionType);
 
     final transactionTypes =
-        result.map((e) => TransactionType.fromMap(e)).toList();
+        result.map((e) => TransactionTypeModel.fromMap(e)).toList();
 
     return transactionTypes;
   }
