+++ lib/shared/usecases/get_transaction_types_to_dropdown_use_case.dart
@@ -1,4 +1,4 @@
-import 'package:hermes_app/shared/entities/transaction_type.dart';
+import 'package:hermes_app/shared/entities/transaction_type_model.dart';
 import 'package:hermes_app/shared/repositories/transaction_type_repository.dart';
 import 'package:hermes_app/shared/widgets/dropdown/dropdown_item.dart';
 
@@ -16,7 +16,8 @@ class GetTransactionTypesToDropdownUseCase {
   }
 }
 
-extension _TransactionTypeToDropdownItemTransformer on List<TransactionType> {
+extension _TransactionTypeToDropdownItemTransformer
+    on List<TransactionTypeModel> {
   List<DropdownItem> transform() {
     return List.from(
       map(
