+++ lib/shared/utils/text_size.dart
@@ -1,0 +1,1 @@
+enum TextSize { small, medium }
