+++ lib/shared/widgets/default_error_widget/register_error_cubit.dart
@@ -1,5 +1,5 @@
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:hermes_app/shared/models/failure.dart';
+import 'package:hermes_app/shared/entities/failure_model.dart';
 
 class RegisterErrorCubit extends Cubit {
   RegisterErrorCubit(super.initialState);
